Learning RSpec; Abbrev::Builder
I'm trying to learn RSpec, so I've written a spec for something I
found wanting in abbrev.rb: an abbreviation builder that supports
incremental addition of words and generates truly unambiguous
abbreviations (that is, no word should get an abbreviation that is a
prefix of any other word).

Here is a spec and the beginnings of an implementation. Once the
implementation is done, I'll refactor it to make it slightly better in
order to glean some experience or refactoring ruby code that does
something interesting under RSpec.

	new file:   .rspec
	new file:   Gemfile
	new file:   Gemfile.lock
	new file:   lib/abbrev_builder.rb
	new file:   lib/abbrev_builder_spec.rb